Create session.py module
